+++ libs/wsnet/src/httpnetworkmanager/httpnetworkmanager.h
@@ -14,16 +14,16 @@ public:
 
     bool init();
 
-    std::shared_ptr<WSNetHttpRequest> createGetRequest(const std::string &url, std::uint16_t timeoutMs,
+    std::shared_ptr<WSNetHttpRequest> createGetRequest(const std::string &url, std::uint32_t timeoutMs,
                                                        bool isIgnoreSslErrors = false) override;
 
-    std::shared_ptr<WSNetHttpRequest> createPostRequest(const std::string &url, std::uint16_t timeoutMs,
+    std::shared_ptr<WSNetHttpRequest> createPostRequest(const std::string &url, std::uint32_t timeoutMs,
                                                         const std::string &data, bool isIgnoreSslErrors = false) override;
 
-    std::shared_ptr<WSNetHttpRequest> createPutRequest(const std::string &url, std::uint16_t timeoutMs,
+    std::shared_ptr<WSNetHttpRequest> createPutRequest(const std::string &url, std::uint32_t timeoutMs,
                                                        const std::string &data, bool isIgnoreSslErrors = false) override;
 
-    std::shared_ptr<WSNetHttpRequest> createDeleteRequest(const std::string &url, std::uint16_t timeoutMs,
+    std::shared_ptr<WSNetHttpRequest> createDeleteRequest(const std::string &url, std::uint32_t timeoutMs,
                                                           bool isIgnoreSslErrors = false) override;
 
     std::shared_ptr<WSNetCancelableCallback> executeRequestEx(const std::shared_ptr<WSNetHttpRequest> &request, std::uint64_t id,
